+++ build.sh
@@ -282,23 +282,28 @@ build_apple()
         if [ "$platform" = "$bitcode_platform" ]; then
             word_list_append "cflags_arch" "-mstrict-align" || exit 1
             word_list_append "cflags_arch" "-m$os_name-version-min=$min_version" || exit 1
-            word_list_append "cflags_arch" "-fembed-bitcode" || exit 1
+            if [ "$enable_bitcode" = "yes" ]; then
+                word_list_append "cflags_arch" "-fembed-bitcode" || exit 1
+            fi
         else
             word_list_append "cflags_arch" "-m$simulator_name-simulator-version-min=$min_version" || exit 1
-            word_list_append "cflags_arch" "-fembed-bitcode-marker" || exit 1
+            if [ "$enable_bitcode" = "yes" ]; then
+                word_list_append "cflags_arch" "-fembed-bitcode-marker" || exit 1
+            fi
         fi
         sdk_root="$xcode_home/Platforms/$platform.platform/Developer/SDKs/$sdk"
-        $MAKE -C "src/realm" "librealm-$platform.a" "librealm-$platform-dbg.a" BASE_DENOM="$platform" CFLAGS_ARCH="$cflags_arch -isysroot '$sdk_root'" || exit 1
-        mkdir "$temp_dir/platforms/$platform" || exit 1
-        cp "src/realm/librealm-$platform.a"     "$temp_dir/platforms/$platform/librealm.a"     || exit 1
-        cp "src/realm/librealm-$platform-dbg.a" "$temp_dir/platforms/$platform/librealm-dbg.a" || exit 1
+        tag="$platform$platform_suffix"
+        $MAKE -C "src/realm" "librealm-$tag.a" "librealm-$tag-dbg.a" BASE_DENOM="$tag" CFLAGS_ARCH="$cflags_arch -isysroot '$sdk_root'" || exit 1
+        mkdir "$temp_dir/platforms/$tag" || exit 1
+        cp "src/realm/librealm-$tag.a"     "$temp_dir/platforms/$tag/librealm.a"     || exit 1
+        cp "src/realm/librealm-$tag-dbg.a" "$temp_dir/platforms/$tag/librealm-dbg.a" || exit 1
     done
     REALM_ENABLE_FAT_BINARIES="1" $MAKE -C "src/realm" "realm-config-$simulator_name" "realm-config-$simulator_name-dbg" BASE_DENOM="$simulator_name" CFLAGS_ARCH="-fembed-bitcode -DREALM_CONFIG_$all_caps_name" AR="libtool" ARFLAGS="-o" || exit 1
     mkdir -p "$dir" || exit 1
-    echo "Creating '$dir/librealm-$simulator_name.a'"
-    libtool "$temp_dir/platforms"/*/"librealm.a"     -static -o "$dir/librealm-$simulator_name.a"     || exit 1
+    echo "Creating '$dir/librealm-$simulator_name$platform_suffix.a'"
+    libtool "$temp_dir/platforms"/*/"librealm.a"     -static -o "$dir/librealm-$simulator_name$platform_suffix.a"     || exit 1
     echo "Creating '$dir/librealm-$simulator_name-dbg.a'"
-    libtool "$temp_dir/platforms"/*/"librealm-dbg.a" -static -o "$dir/librealm-$simulator_name-dbg.a" || exit 1
+    libtool "$temp_dir/platforms"/*/"librealm-dbg.a" -static -o "$dir/librealm-$simulator_name$platform_suffix-dbg.a" || exit 1
     echo "Copying headers to '$dir/include'"
     mkdir -p "$dir/include" || exit 1
     cp "src/realm.hpp" "$dir/include/" || exit 1
@@ -311,6 +316,7 @@ build_apple()
         y="$(printf "%s\n" "$x" | sed "s/realm-config/realm-config-$simulator_name/")" || exit 1
         cp "src/realm/$y" "$REALM_HOME/$dir/$x" || exit 1
     done
+    rm -rf "$temp_dir"
     echo "Done building"
     return 0
 }
@@ -764,6 +770,12 @@ EOF
         export sdks_config_key='IPHONE_SDKS'
         export dir="$IPHONE_DIR"
         export all_caps_name='IOS'
+        export platform_suffix='-bitcode'
+        export enable_bitcode='yes'
+        build_apple
+
+        export platform_suffix='-no-bitcode'
+        export enable_bitcode='no'
         build_apple
         ;;
 
@@ -778,6 +790,8 @@ EOF
         export sdks_config_key='WATCHOS_SDKS'
         export dir="$WATCHOS_DIR"
         export all_caps_name='WATCHOS'
+        export platform_suffix=''
+        export enable_bitcode='yes'
         build_apple
         ;;
 
@@ -792,6 +806,8 @@ EOF
         export sdks_config_key='TVOS_SDKS'
         export dir="$TVOS_DIR"
         export all_caps_name='TVOS'
+        export platform_suffix=''
+        export enable_bitcode='yes'
         build_apple
         ;;
 
@@ -981,14 +997,7 @@ EOF
         rm -f "$BASENAME-$realm_version.zip" || exit 1
         mkdir -p "$tmpdir/$BASENAME/include" || exit 1
         cp -r "$IPHONE_DIR/include/"* "$tmpdir/$BASENAME/include" || exit 1
-        for original in "$IPHONE_DIR/librealm-ios.a" "$IPHONE_DIR/librealm-ios-dbg.a" "$WATCHOS_DIR/librealm-watchos.a" "$WATCHOS_DIR/librealm-watchos-dbg.a" "src/realm/librealm.a" "src/realm/librealm-dbg.a"; do
-            cp "$original" "$tmpdir/$BASENAME" || exit 1
-        done
-        for sdk in $iphone_sdks $watchos_sdks; do
-            platform="$(printf "%s\n" "$sdk" | cut -d: -f1)" || exit 1
-            cp "src/realm/librealm-$platform.a" "$tmpdir/$BASENAME" || exit 1
-            cp "src/realm/librealm-$platform-dbg.a" "$tmpdir/$BASENAME" || exit 1
-        done
+        cp "$IPHONE_DIR"/*.a "$WATCHOS_DIR"/*.a "src/realm/librealm.a" "src/realm/librealm-dbg.a" "$tmpdir/$BASENAME" || exit 1
         cp tools/LICENSE "$tmpdir/$BASENAME" || exit 1
         if ! [ "$REALM_DISABLE_MARKDOWN_CONVERT" ]; then
             command -v pandoc >/dev/null 2>&1 || { echo "Pandoc is required but it's not installed.  Aborting." >&2; exit 1; }